backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
sqlalchemy>=2.0
asyncpg
alembic
python-dotenv
httpx
opencv-python
httpx
Pillow
openai
numpy
pytest
pytest-mock
pytest-cov
testcontainers[postgres]
psycopg2-binary